+++ src/CleanRegex/Exception/ExplicitDelimiterRequiredException.php
@@ -1,7 +1,7 @@
 <?php
 namespace TRegx\CleanRegex\Exception;
 
-class ExplicitDelimiterRequiredException extends \Exception implements PatternException
+class ExplicitDelimiterRequiredException extends \RuntimeException implements PatternException
 {
     public function __construct(string $message)
     {