+++ src/CleanRegex/Exception/InternalCleanRegexException.php
@@ -1,6 +1,6 @@
 <?php
 namespace TRegx\CleanRegex\Exception;
 
-class InternalCleanRegexException extends \Exception implements PatternException
+class InternalCleanRegexException extends \RuntimeException implements PatternException
 {
 }