+++ src/CleanRegex/Exception/InvalidIntegerTypeException.php
@@ -3,7 +3,7 @@ namespace TRegx\CleanRegex\Exception;
 
 use TRegx\CleanRegex\Internal\Type\Type;
 
-class InvalidIntegerTypeException extends \Exception implements PatternException
+class InvalidIntegerTypeException extends \RuntimeException implements PatternException
 {
     public static function forInvalidType(Type $type): self
     {