+++ src/CleanRegex/Exception/NoSuchStreamElementException.php
@@ -3,7 +3,7 @@ namespace TRegx\CleanRegex\Exception;
 
 use TRegx\CleanRegex\Internal\Message\Message;
 
-class NoSuchStreamElementException extends \Exception implements PatternException
+class NoSuchStreamElementException extends \RuntimeException implements PatternException
 {
     public function __construct(Message $message)
     {