+++ src/CleanRegex/Exception/NonexistentGroupException.php
@@ -3,7 +3,7 @@ namespace TRegx\CleanRegex\Exception;
 
 use TRegx\CleanRegex\Internal\GroupKey\GroupKey;
 
-class NonexistentGroupException extends \Exception implements PatternException
+class NonexistentGroupException extends \RuntimeException implements PatternException
 {
     public function __construct(GroupKey $group)
     {